Fix a crash in the shbench fibonacci benchmark
If the fibslow function in the shbench[*] fibonacci benchmark
is run with an argument greater than 20, ksh will crash with
a memory fault. This was caused by subshell numbers only being
16-bits long, which isn't enough when running a large number
of subshells. Extending the subshell number length to 64-bit
fixes this issue, but with vmalloc only to a certain extent.
The fibonacci benchmark no longer crashes when run, but the
regression test this commit adds only works with standard malloc.
With ksh93u+ vmalloc the regression test will crash.

src/cmd/ksh93/include/defs.h,
src/cmd/ksh93/include/jobs.h,
src/cmd/ksh93/sh/jobs.c,
src/cmd/ksh93/sh/subshell.c,
- Extend curenv, jobenv, subenv and p_env from short to long.
  This bugfix was backported from ksh93v- with a patch from
  OpenSUSE: https://build.opensuse.org/package/view_file/shells/ksh/ksh93-longenv.dif?expand=1

src/cmd/ksh93/tests/subshell.sh:
- Add a regression test for running a large number of
  subshells, but disable it if vmstate is a builtin (since
  it doesn't work with vmalloc).

[*]: https://github.com/ksh-community/shbench

+++ src/cmd/ksh93/sh/subshell.c
@@ -103,7 +103,7 @@ static struct subshell
 	char		pwdclose;
 } *subshell_data;
 
-static int subenv;
+static long subenv;
 
 
 /*
@@ -176,7 +176,7 @@ void sh_subfork(void)
 {
 	register struct subshell *sp = subshell_data;
 	Shell_t	*shp = sp->shp;
-	int	curenv = shp->curenv;
+	long curenv = shp->curenv;
 	pid_t pid;
 	char *trap = shp->st.trapcom[0];
 	if(trap)
@@ -456,7 +456,7 @@ Sfio_t *sh_subshell(Shell_t *shp,Shnode_t *t, volatile int flags, int comsub)
 	struct subshell sub_data;
 	register struct subshell *sp = &sub_data;
 	int jmpval,nsig=0,duped=0;
-	int savecurenv = shp->curenv;
+	long savecurenv = shp->curenv;
 	int savejobpgid = job.curpgid;
 	int *saveexitval = job.exitval;
 	char *savsig;
